requirements.txt
gradio
python-dotenv
pytest
pytest-mock
pytest-cov
strawberry-graphql
supabase
